app/legacy_db/customer_read.py
# Archived: app/services/customer_read.py
from sqlalchemy import text
from sqlalchemy.ext.asyncio import AsyncSession

async def get_customer_profile(session: AsyncSession, customer_id: int):
    # archived
    return None
